+++ src/main/java/com/ureca/uble/domain/bookmark/controller/BookmarkController.java
@@ -4,8 +4,8 @@ import com.ureca.uble.domain.bookmark.dto.response.CreateBookmarkRes;
 import com.ureca.uble.domain.bookmark.dto.response.DeleteBookmarkRes;
 import com.ureca.uble.domain.bookmark.dto.response.GetBookmarkRes;
 import com.ureca.uble.domain.bookmark.service.BookmarkService;
-import com.ureca.uble.global.response.CommonResponse;
-import com.ureca.uble.global.response.CursorPageRes;
+import com.ureca.uble.domain.common.dto.response.CommonResponse;
+import com.ureca.uble.domain.common.dto.response.CursorPageRes;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
