+++ src/main/java/com/ureca/uble/domain/brand/controller/BrandController.java
@@ -6,8 +6,8 @@ import com.ureca.uble.domain.brand.dto.response.SearchBrandListRes;
 import com.ureca.uble.domain.brand.service.BrandService;
 import com.ureca.uble.entity.enums.BenefitType;
 import com.ureca.uble.entity.enums.Season;
-import com.ureca.uble.global.response.CommonResponse;
-import com.ureca.uble.global.response.CursorPageRes;
+import com.ureca.uble.domain.common.dto.response.CommonResponse;
+import com.ureca.uble.domain.common.dto.response.CursorPageRes;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
