+++ src/main/java/com/ureca/uble/domain/feedback/controller/FeedbackController.java
@@ -4,13 +4,11 @@ import com.ureca.uble.domain.feedback.dto.request.CreateFeedbackReq;
 import com.ureca.uble.domain.feedback.dto.response.AdminFeedbackRes;
 import com.ureca.uble.domain.feedback.dto.response.CreateFeedbackRes;
 import com.ureca.uble.domain.feedback.service.FeedbackService;
-import com.ureca.uble.global.response.CommonResponse;
+import com.ureca.uble.domain.common.dto.response.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
