+++ src/main/java/com/ureca/uble/domain/users/controller/UsageHistoryController.java
@@ -4,8 +4,8 @@ import com.ureca.uble.domain.users.dto.request.CreateUsageHistoryReq;
 import com.ureca.uble.domain.users.dto.response.CreateUsageHistoryRes;
 import com.ureca.uble.domain.users.dto.response.UsageHistoryRes;
 import com.ureca.uble.domain.users.service.UsageHistoryService;
-import com.ureca.uble.global.response.CommonResponse;
-import com.ureca.uble.global.response.CursorPageRes;
+import com.ureca.uble.domain.common.dto.response.CommonResponse;
+import com.ureca.uble.domain.common.dto.response.CursorPageRes;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
