+++ src/main/java/com/ureca/uble/global/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.ureca.uble.global.exception;
 
-import com.ureca.uble.global.response.CommonResponse;
+import com.ureca.uble.domain.common.dto.response.CommonResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
